requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.7
lightning-utilities==0.15.2
markupsafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt-toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.4
pyyaml==6.0.2
pyzmq==27.0.2
scallopy @ file:///home/ehrenhoefer/playground/scallopy-0.1.4-cp39-cp39-manylinux_2_27_x86_64.whl
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typing-extensions==4.15.0
wcwidth==0.2.13
yarl==1.20.1
zipp==3.23.0
